+++ SDL_geometry.h
@@ -20,7 +20,7 @@ typedef struct GEOM_Vertex
 } GEOM_Vertex;
 
 int GEOM_RenderGeometry(SDL_Renderer *renderer, SDL_Texture *texture,
-                        GEOM_Vertex *vertices, int num_vertices, int *indices,
+                        GEOM_Vertex *vertices, int num_vertices, Uint16 *indices,
                         int num_indices);
 
 #ifdef __cplusplus
